requirements.txt
fastapi
mediapipe
opencv-python
numpy
python-multipart
uvicorn
cloudinary
python-dotenv